Add an `XHPAST::getNamespace()` method
Summary: This throws away D14498 and starts again. Basically, the grammar of PHP does not allow `namespace` statements to be parsed consistently (see some discussion in D14506). Instead of complicating the grammar, move the logic of determining a node's namespace to the PHP representation of the AST.

Test Plan: Added unit tests.

Reviewers: richardvanvelzen, #blessed_reviewers, epriestley

Reviewed By: #blessed_reviewers, epriestley

Subscribers: Korvin

Differential Revision: https://secure.phabricator.com/D14512