Improve parsing of `use` statements
Summary:
The way that we parse `use` statements at the moment is a bit odd. For example, consider the following example:

```lang=php
<?php

use A\B\C;
use X\Y\Z;
```

This is currently parsed as:

{F976619, size=full}

After this change, it is instead parsed as:

{F976621, size=full}

Test Plan: Added a test case.

Reviewers: epriestley, #blessed_reviewers

Reviewed By: epriestley, #blessed_reviewers

Subscribers: Korvin

Differential Revision: https://secure.phabricator.com/D14518